New feature added: Configuration synchronization between proxy and backend.
Objective:
being able to consume the centralized configuration in the proxy from the servers, allowing only the proxy to be used as a database credentials manager.
Added packet encryption with AES/GCM/NoPadding protocol.

Consequence:
-Change plugin initialization flow to manage delayed initialization without causing dependency conflicts.
-Massive refactoring of static references and injectable components that hinder the architecture in general, preventing the initialization flow orchestrated by the engine from being carried out.

first commit with the system tested and working as expected in "velocity", "paper","folia".

+++ Spigot/src/main/java/BlockDynasty/BukkitImplementation/Integrations/treasuryEconomy/EconomyHook.java
@@ -28,6 +28,7 @@ import me.lokka30.treasury.api.economy.account.AccountPermission;
 import me.lokka30.treasury.api.economy.account.NonPlayerAccount;
 import me.lokka30.treasury.api.economy.account.accessor.AccountAccessor;
 import me.lokka30.treasury.api.economy.currency.Currency;
+import net.blockdynasty.providers.services.ServiceProvider;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
@@ -36,19 +37,27 @@ import java.util.*;
 import java.util.concurrent.CompletableFuture;
 
 public class EconomyHook implements EconomyProvider {
-    private final DynastyEconomy api;
 
-    public EconomyHook(DynastyEconomy api) {
-       this.api = api;
+    public EconomyHook() {
+
     }
+    private Optional<DynastyEconomy> getApi(){
+        return ServiceProvider.get(DynastyEconomy.class, service -> service.getId().equals(com.blockdynasty.economy.Economy.getApiWithVaultLoggerId()));
+    }
+
 
     @Override
     public @NotNull AccountAccessor accountAccessor() {
-        return new AccountTreasury(api);
+        return new AccountTreasury();
     }
 
     @Override
     public @NotNull CompletableFuture<Boolean> hasAccount(@NotNull AccountData accountData) {
+        Optional<DynastyEconomy> optionalApi = getApi();
+        if (optionalApi.isEmpty()) {
+            return CompletableFuture.completedFuture(false);
+        }
+        DynastyEconomy api = optionalApi.get();
         Optional<UUID> optional =accountData.getPlayerIdentifier();
         if (optional.isPresent()) {
             UUID playerId = optional.get();
@@ -62,12 +71,22 @@ public class EconomyHook implements EconomyProvider {
 
     @Override
     public @NotNull CompletableFuture<Collection<UUID>> retrievePlayerAccountIds() {
+        Optional<DynastyEconomy> optionalApi = getApi();
+        if (optionalApi.isEmpty()) {
+            return CompletableFuture.completedFuture(Collections.emptyList());
+        }
+        DynastyEconomy api = optionalApi.get();
         List<Account> accounts = api.getAccountsOffline();
         return CompletableFuture.completedFuture(accounts.stream().map(Account::getUuid).toList());
     }
 
     @Override
     public @NotNull CompletableFuture<Collection<NamespacedKey>> retrieveNonPlayerAccountIds() {
+        Optional<DynastyEconomy> optionalApi = getApi();
+        if (optionalApi.isEmpty()) {
+            return CompletableFuture.completedFuture(Collections.emptyList());
+        }
+        DynastyEconomy api = optionalApi.get();
         List<Account> accounts = api.getAccountsOffline();
         List<String> list = accounts.stream().map(Account::getName).toList();
         return CompletableFuture.completedFuture(list.stream().map(name -> NamespacedKey.of("blockdynasty",name)).toList());
@@ -85,11 +104,20 @@ public class EconomyHook implements EconomyProvider {
 
     @Override
     public @NotNull Currency getPrimaryCurrency() {
-        return new CurrencyTreasury(api.getDefaultCurrency());
+        Optional<DynastyEconomy> optionalApi = getApi();
+        if (optionalApi.isEmpty()) {
+            return new CurrencyTreasury(null);
+        }
+        return new CurrencyTreasury(optionalApi.get().getDefaultCurrency());
     }
 
     @Override
     public @NotNull Optional<Currency> findCurrency(@NotNull String identifier) {
+        Optional<DynastyEconomy> optionalApi = getApi();
+        if (optionalApi.isEmpty()) {
+            return Optional.empty();
+        }
+        DynastyEconomy api = optionalApi.get();
         com.BlockDynasty.api.entity.Currency c = api.getCurrency(identifier);
         if (c == null) {
             return Optional.empty();
@@ -104,6 +132,11 @@ public class EconomyHook implements EconomyProvider {
 
     @Override
     public @NotNull Set<Currency> getCurrencies() {
+        Optional<DynastyEconomy> optionalApi = getApi();
+        if (optionalApi.isEmpty()) {
+            return Set.of();
+        }
+        DynastyEconomy api = optionalApi.get();
         List< com.BlockDynasty.api.entity.Currency> c= api.getCurrencies();
         if (c != null) {
             Set<Currency> currencies = new HashSet<>();
@@ -122,6 +155,11 @@ public class EconomyHook implements EconomyProvider {
 
     @Override
     public @NotNull CompletableFuture<TriState> registerCurrency(@NotNull Currency currency) {
+        Optional<DynastyEconomy> optionalApi = getApi();
+        if (optionalApi.isEmpty()) {
+            return CompletableFuture.completedFuture(TriState.FALSE);
+        }
+        DynastyEconomy api = optionalApi.get();
         EconomyResponse response = api.createCurrency(currency.getDisplayName(BigDecimal.ZERO,Locale.getDefault()),currency.getDisplayName(BigDecimal.ZERO,Locale.getDefault()));
         if (response.isSuccess()){
             return CompletableFuture.completedFuture(TriState.TRUE);
@@ -132,6 +170,11 @@ public class EconomyHook implements EconomyProvider {
 
     @Override
     public @NotNull CompletableFuture<TriState> unregisterCurrency(@NotNull Currency currency) {
+        Optional<DynastyEconomy> optionalApi = getApi();
+        if (optionalApi.isEmpty()) {
+            return CompletableFuture.completedFuture(TriState.FALSE);
+        }
+        DynastyEconomy api = optionalApi.get();
         EconomyResponse response = api.deleteCurrency(currency.getDisplayName(BigDecimal.ZERO,Locale.getDefault()));
         if (response.isSuccess()){
             return CompletableFuture.completedFuture(TriState.TRUE);
